backend/shop/shop/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
python-memcached==1.62
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
gunicorn==22.0.0